requirements.txt
aiohttp
aiosignal
annotated-types
anyio
asttokens
async-timeout
attrs
certifi
charset-normalizer
colorama
comm
dataclasses-json
debugpy
decorator
distro
exceptiongroup
executing
frozenlist
greenlet
h11
httpcore
httpx
idna
ipykernel
ipython
jedi
jsonpatch
jsonpointer
jupyter_client
jupyter_core
langchain
langchain-community
langchain-core
langchain-openai
langsmith
marshmallow
matplotlib-inline
multidict
mypy-extensions
nest-asyncio
numpy
openai
packaging
parso
platformdirs
prompt-toolkit
psutil
pure-eval
pydantic
pydantic_core
Pygments
PyYAML
pyzmq
regex
requests
six
sniffio
SQLAlchemy
stack-data
tenacity
tiktoken
tornado
tqdm
traitlets
typing-inspect
typing_extensions
urllib3
wcwidth
yarl
wget
streamlit
langchain
langchain-community
langchain-google-genai
python-dotenv
google-generativeai